feat: Add LoggerPort to formalize logger dependency (ADR-005)
Introduces LoggerPort protocol in domain/ports.py to abstract logging
dependency, consistent with other ports (MetricsPort, StoragePort, etc.).
Removes direct structlog.BoundLogger coupling from PipelineServices.

- Add LoggerPort Protocol with bind, info, warning, error, debug, exception
- Update PipelineServices to use LoggerPort instead of structlog.BoundLogger
- Update PipelineContext to use LoggerPort
- Keep BoundLogger as backward compatibility alias in context.py
- Document decision in ADR-005-logger-metrics-ports.md
